Split each check out into a separate function
This allows us to easily run selected checks only.
This also splits the check logic up nicely so that it can be
refactored without affecting the others.

Common tasks such as tokenising the incoming data or finding the starting
tokens of classes has been pulled out of the check logic as well to DRY
things up.