Allow double quoted strings that use backslash
This is a kind of janky solution as Ruby automatically detects the use of the
backslash and strips it out before I can do anything with it, so what we
attempt to do now is look for a block of spaces the same length as the column
number of the start of the :STRING token (the indent) and replace it with
newline character.  The downside of this is that in order for this check to
not throw a false positive, the split string *MUST* be indented exactly right.

Closes #51