Allow lines with only whitespace
Currently the trailing_whitespace check flags a line with only
whitespace (ie. as inserted by most code focussed text editors when
leaving a blank line within an indented block of code).
This allows lines that have only space chars (regex \s) between the
start and end. Not sure if this is the most elegant way to achieve
this, I don't know the puppet-lint tree that well.