Pass all enclosed values in strs to the lexer for tokenising
Previously, a seperate logic block would handle enclosed variables and
anything else would be passed to a new instance of the lexer for
tokenisation. Unfortunately, this lead to variables that had array or hash
references being classified as :NAME tokens and then when running with fix
mode enabled, would not be turned back into variables.