Added --relative command line argument.
This argument will cause the autoload layout class plugin to ignore the
top-most directory in the structure. This is useful for the common case where
the folder name is not the same as the module name in raw source code.

Example:
puppet-foo/manifests/init.pp has 'class foo { }'
puppet-lint puppet-foo returns errors
puppet-lint --relative puppet-foo returns no errors

This allows us to continue testing the general autoload layout if we purposely
compare relatively.