fix #331 - clear any pre-(auto-)existing tasks
Because we automatically create a lint task at the bottom of the file
using a configuration block with the default name (:lint) would not
work, as we would now have 2 instances of that RakeTask floating around.
However, only the first - auto-generated - task is used. Which makes
it impossible to configure any of the options for it.