Ensure that default cases on nested case statements get matched up correctly
Previously, it was possible to miss case statements without a default
case when there were nested case statements in the mix.

This commit fixes that problem by ensuring that default cases are
matched up with their owning case statement, and not with any other
case statements that might exist within the scope.