Add in an option to set arbitrary variables as default scope.
The use case for this is a bit strange, but if you add some variable in
the default scope (as my company has done), you will trigger several
erroneous lint warnings as those variables don't specify the top-scope
(::serverenv). The check, however, is still valid in several places, and
we don't want to not trigger on that warning when it does detect a
legitimate error.

This changeset will add in an optional parameter to the puppet-lint CLI
and Rake tasks that allow you to specify a list of variables that will
be ignored from this check.

Tests were added and the README was updated in this commit.