Add in the ability to specify custom puppet:// fileserver urls.
This option takes in a filepath as a parameter. It does some very naive
stripping of commented strings and matches against lines with the
pattern /\[(\w+)\], such as [bigfiles]. It takes the captured group and
then does the same check that was done for the 'module' string.

This code is a bit hairier than my last PR, and I'm sure it can be done
better. I know there were plans to split this check off into a community
plugin, and I would gladly apply this there if that ever becomes the
case.