Sidesteps a crash when running puppet-lint -f in some situations where the internal calculation of tab-width in lib/puppet-lint/plugins/check_whitespace.rb's fix(problem) ends up with a negative number.
The crash occurs on a line which looks like:

      ensure=>"purged"

In the context of:

  if $removeiceweasel == true {
    package{"removeiceweasel":
      name=>"iceweasel",
      ensure=>"purged"
    }
  }

When the file is of type:

    ASCII text, with CRLF line terminators

Specifically in file pkg/runthebusiness-firefox-1.0.0/manifests/init.pp on commit ff3fc7a of the repo https://github.com/runthebusiness/puppet-firefox.

Changing the line terminators from \r\n to \n does not solve the issue.