This adds a disabled-by-default 80chars check
This commit adds back an 80chars check, but disables it by default. To
do this, I also added the ability to re-enable checks that have been
disabled.

The spec is failing and I don't have time to fix it just now, but it
works when called from the command line. I'll come back to it if nobody
else picks this up.

A better solution than calling `Puppet.configuration` to disable a check
like I did might be to extend the `new_check` method. That's not
orthagonal to this PR though.

Supercedes #488
Fixes #484