Use lookahead assertion for matching function name
For some reason using a lookahead assertion for the opening parenthesis
after a function name is much faster for large manifests. The old
version shows some pathologic behaviour and becomes very slow for
manifests with many entries -- as described in #717.

I don't understand why the old version behaves badly. I was able to
reproduce the speedup with Ruby 1.9.3 as well as with 2.3.something.