Include check name in log output by default
As reported in https://github.com/rodjek/puppet-lint/issues/837 it's
sometimes difficult to identify the check that causes a particular
warning or error output. By including the check name in the log output
it becomes easier to disable checks via control comments or command
line parameters. The old format can be restored with '--log-format'.