(#794) Dockerfile for puppet-lint
Before we begin, it should be noted that I don't regularly use Docker so
there may be some bad practices here. If so, please let me know.

This Dockerfile is pretty minimal and creates a working image of ~21MB
which I think is pretty good (5MB of that is alpine and 15MB is the Ruby
runtime).

The files to be linted should be mounted to the `/puppet` volume, e.g.
linting the manifests in my current directory:
```
$ docker run --rm -v $PWD:/puppet rodjek/puppet-lint *.pp
866.pp - ERROR: gh::issue866 not in autoload module layout on line 2
869.pp - ERROR: profile::users not in autoload module layout on line 2
869.pp - ERROR: profile::users::geco not in autoload module layout on line 22
```

If this seems OK to the interested parties, I'll merge this and then set
up automatic build & push on tag to Docker Hub.