Do not rely on tokens array positioning when fixing ensure_first_param problems
Currently the fix method for this check relies upon the position of the
first parameter in the tokens array to operate which is determined when
the manifest is first loaded. If checks with fix methods that add or
remove tokens (like the `trailing_whitespace` check) run before this
check, the saved position becomes no longer accurate causing the fix
method to fail.

Fixes #906
Fixes #905
Fixes #839