(MAINT) Rubocop fixes and refactor load_from_gems
Previously load_from_gems contained nested loops and block
continuation. It was hard to read and probably not that efficient when
processing many files.

This commit cleans the method up by splitting the operations out in to
three distinct operations, hopefully simplifying the code slightly.

In addition, the Rubocop fixes have also been applied in this commit.