(MAINT) Optional refactor of line length check
Prior to this commit there was some code duplication between the
140char and 80char checks.

This PR breaks out the core logic for these checks in to it's own class.
This reduces duplication and also allows for a configurable line length.

In the future this could be expanded to allow the user to specify their
desired line length from config or a flag.