(MAINT) Restructure spec files
This commit restructures the files in spec. It moves all existing tests
in to a new folder called unit. This is to make way for acceptance tests
that will live in a folder called acceptance.

Additionally the Rakefile has been updated to exclude files that are not
unit tests for the `spec` task.