(MAINT) Add spec helpers
This commit adds two spec helpers that enable litmus and provide
additional functionality.

For example:

* A custom runner has been created to execute puppet-lint via it's
entrypoint. The function will also ensure that the resulting command
is executed in a safe manor.

* Two new rspec matchers have been added that help verify the messages
that puppet-lint returns. They are very similar to the custom matchers
that exist for the spec test, just much simpler.