(MAINT) Fix spec test paths
Prior to this commit a number of spec tests were failing. This was
because tests were expecting files to exist in the previous location.

Some could not load `spec_helper` and others were referencing a path
that no longer existed.

This commit fixes the rspec filter in `spec_helper` so that it properly
includes all of the unit tests and updates the expected paths in
`bin_spec`.