(MAINT) Fix rubocop inheritance
Prior to this commit we attempted to configure the puppet-lint gem so
downstream plugin authors could consume rubocops config via the gem.

This mostly worked however there seemed to be an issue when inheriting
`AllCops/Exclude` properties.

tldr; inheriting a config called `.rubocop.yml` doesn't seem to work as
expected.

This commit attempts to solve that by splitting out the main rules for
puppet-lint in to a new file called `rubocop_baseline.yml`. Them,
.rubocop.yml has been configured to inherit from this file.

This means that puppet-lint can continue to operate as normal.

The `gemspec` entry for `.rubocop.yml` has been replaced with
`rubocop_baseline.yml`. This ensures that we are packaging the baselines
config with the puppet-lint gem.

Downstream plugins can then inherit this configuration like this:

```
inherit_gem:
  puppet-lint:
    - rubocop_baseline.yml
```

This appears to pull in the `AllCops/Exclude` properties as expected.