Fix assertion when rspec-json_expectations is missing
In 21f5743521fe36f4cc04f87ea62b63d9b5a9567e the output was changed and
the code path when rspec-json_expectations is present was updated, but
not without. While this is an odd code path (since Gemfile does declare
it as a dependency), it can be hit if a different dependency mechanism
is used and the expectations should be correct.

Fixes: 21f5743521fe36f4cc04f87ea62b63d9b5a9567e