(Bugfix) Fix failing legacy fact autocorrector
Prior to this commit, attempting to autocorrect a legacy fact would
output the wrong result. This happened due to a conflict between the
actions of the legacy fact plugin and the interference of the top
scope fact plugin.

This commit aims to prevent the top scope fact plugin from
attempting to correct any legacy fact by whitelisting all legacy
keywords. Legacy fact plugin already performs the work of the top scope
fact plugin so no functionality should be lost.